Split mascon controller tests
